refactor(tvix): Limit to only 1 thread
Since upstream nix forks a process per connection, there's a bunch of
global, static builder state that is not properly locked or made
thread-local that is causing problems with having more than one builder.
In this particular case, there's a global static flag for whether we're
currently building something to prevent accidental recursive builds
that *aborts* (:0) in the case that the invariant is broken - obviously
we'd like to get rid of that at some point, but in the meantime to get
us to a point where tvix is actually usable let's just limit to 1 thread
at a time. This will cause a bit of a perf hit, but it will hopefully
get us closer to self-hosting.

Change-Id: Ic8460276769491b287e6b594984baae3b66beaa8