Add step to my finances playbook
This is a helpful reminder to ensure that the exchange rate is always
fresh. Ideally I could use Google Sheets formulae to take a snapshot of the
GBP:USD exchange rate *at a point in time*, but I haven't set that up yet. Maybe
it'd look something like:

```pseudocode
=GOOGLEFINANCE("GBP:USD", "September 1, 2020")
```

I'll have to look into this.