chore(tvix): avoid unnesscary rebuilds of protos and tvix
For the build of tvix we can safely ignore all the files in src/protos
as well as all the nix files in the root of the repository. By passing
the input sources through builtins.filterSource we can make a per-file
decision whether or not we want to copy it into the store (in other
words: if we consider it relevant for the build).

This enables much faster development on the actual build expression and
the passed through test-vm expression.

Change-Id: Ib4c01dbe9cbfa9770922f6257af2a7259814ce0d