chore(3p): move quassel to the right inherit statement
As the comment above the inherit statement says it is supposed to only
contain packages required for //third_party/nix. As much as I like IRC I
can't find a usage of Quassel within Nix itself. Perhaps a plugin to
report build status at some point?

Until that plugin mainfests we might as well move it to the correct
location.

Change-Id: I14fd6b698bb779b80829e7bebd5daa69cacdde78