feat(tvix): Support actually building something in the test-vm
The test-vm had the default memory (512MB) and disk size (512MB) limits
which were not particular helpful with building anything from depot.
Unpacking the nixpkgs tarball already failed as / (which isn't on tmpfs)
did run out of inodes. The R/W part of the store is stored in RAM
leading to frequent out of memory situations when unpacking sources
actually succeeded.

In order to actually fetch some of our dependencies we need git on PATH
as builtin fetchers (such as builtins.fetchGit, builtins.fetchtarball,
…) do not use FFI but shell out the respective binaries.

Change-Id: Iaa094646a3cf124900228083f433a8b4daf17f3d