fix(nix/emptyDerivation): always call as a function
The idea with using `__functor` was that it could be called both as a
function and used as a derivation. But it turns out that once you set
`__functor`, nix will always interpret it as a function, so it doesn’t
make sense to keep it around. Apparently we never used it without
overriding some of the fields anyway.

Change-Id: I715ac3f3a0914809b7316b88310c47a0594dae44