Finish Tree section of LC problems
Wahoo! I need to remember that the inorder traversal of a BST should be
sorted. This piece of trivia comes in handy for a variety of BST related
problems.

I also think being able to do a {pre,in,post}-order traversal recursively and
iteratively is a skill that I need to develop.