feat(third_party): separate exposed list to allow more owners
The exposed package list has to be changed/amended quite frequently,
every time somebody wants to use a package not yet in that list and
thus has to whitelist it here.

This effectively requires a superowner review every single time, which
is an unreasonable blocker for many CLs.

I thus propose moving the list into a separate subdirectory and
letting anybody add themselves to the OWNERS of that directory.

(A quick glance at the OWNERS file plugin docs has not showed an
obvious way of auto-allowing anybody).

Change-Id: Ied8bac066e4b9a91ddd642db805fe33dc37872c9