feat(todolist): use static slapd user data for knownUsers
To access the slapd data we simply utilize the data used to generate its
static LDIF data which is checked into depot. This requires moving the
users list out of ops/nixos/tvl-slapd/default.nix into a separate nix
expression, so we can conveniently import it without evaluating a
hypothetical nixos configuration module.

One question is if maybe ops/nixos/tvl-slapd/users.nix should be moved
to a more apt location as it is essentially a general purpose user
registry in depot with this change.

Change-Id: I35caaaab70a5578c47cedc7f33077dd513766290