Revert "feat(nix/readTree): allow arguments to depend on tree location"
This feature isn’t actually used in depot, and it complicates the
logic further, for dubious benefit.

Its semantics and intended use are not documented (neither in code,
nor in test, nor in the original commit message, so I can’t say for
sure, but it’s seems to be something about repository-wide injection
of dependencies.

This reverts commit 58e3608e258d35e1a8b132cd65a20d2200f01b9f.

Change-Id: I5ead308d4aaca9b316743a7571c2aec40e581e96