refact(nix/readTree): refactor code to be more declarative Pt. 1
The `readTree` function handles a lot of cases, as shown by the
documentation and the tests, but the code itself is not very
declarative at the moment.

This is step 0.5 towards making it so, aiming towards moving all
import logic into a separate function which just exports a sum type of
the different states a directory can be in and what it means for
`readTree`.

`readDirVisible` should now export all of the skipping behaviour of
the different combinations of `.skip-subtree` and `default.nix`.

`read-tree` is mostly kept as-is, but should get a lot simpler after
pulling more logic into the match statements.

The `match` function is copied from the (not yet merged) `nix.tags`
module, since the idea of `readTree` is to only depend on nix
builtins.

Change-Id: Ie466dc037b96b1bda6f99e703e816adbaa1694c8