feat(rules.pl): check length of commit message header
As per CONTRIBUTING.md we require the first line of commit messages
to be less than 68 characters in length. This wasn't checked and thus
we have some commit in the repository not conforming to this policy.

This commits add this to the conformant commit message check in gerrit
which should prevent such commits in the future.

Change-Id: I3a4e84cc7f4c281d055269ce15b56928ba9c10aa