refactor(runTestsuite): specialized asserts as cases of assertBool
This change aims to reduce duplication of the assert logic by
implementing the generic assertion function assertBool while
preserving user friendly error messages and making it easy to implement
new types of asserts.

This is achieved by restricting the AssertResult sum type back to the
two cases yep and nope and passing the error information as
AssertErrorContext in the context attribute of nope.

The generic assert is implemented as assertBoolContext which succeeds
if a given boolean is true and returns a given AssertErrorContext if
the boolean is false. This allows us to implement assertThrows and
assertEq in terms of assertBoolContext without sacrifizing any error
information.

We also add the user facing assertBool which is trivial to implement.

Change-Id: Ifd6b3aa4187c90c3add2df63fa7c906c8f03fd2d