feat(sterni/ncla-run): shebang interpreter for NCLA™ programs
ncla-run is a tiny wrapper around nix-instantiate which allows to run
NCLA™ (Nix Command Line Application) programs, a novel interface for
writing purely function command line programs.

A NCLA™ program is a nix file which conforms to the following
constraints:

* It must evaluate to a function which has a set pattern with an
  ellipsis as the single argument.

* It must produce a string as a return value or fail.

When invoked, a NCLA™ program receives the following arguments:

* `currentDir`: the current working directory as a nix path

* `argv`: a list of strings containing `argv` including `argv[0]`

* extra arguments which are manually specified which allows for
  passing a long dependencies or libraries, for example:

    ncla-run --arg depot '(import /depot {})' my-prog.nix [ argv[1] … ]

  would pass along depot to be used in `my-prog.nix`.

The NCLA™ purity derives from the fact that such nix expressions can
cause effects only in a very narrow sense (using builtins.fetchurl
if TARBALL_TTL is 0, adding files and directories to the nix store,
realising derivations). As an approximation, a NCLA™ program can be
thought of a function with the following signature:

  λ :: environment → working directory → argv → stdout

where environment includes:

* the time at the start of the program (`builtins.currentTime`)
* other information about the machine (`builtins.currentSystem` …)
* environment variables (`builtins.getEnv`)
* the file system (`builtins.readDir`, `builtins.readFile`, …) which
  is the biggest input impurity as it may change during evaluation

Additionally import from derivation and builtin fetchers are available
which introduce further impurities to be utilized.

Future work:

* Streaming I/O via lazy lists. This would allow usage of
  stdin and output before the program terminates. However this would
  require using libexpr directly or writing a custom nix interpreter.

  A description of how this would work can be found on the website of the
  esoteric programming language Lazy K: https://tromp.github.io/cl/lazy-k.html

* An effect system beyond stdin / stdout.

* Better error handling, support setting exit codes etc.

Change-Id: I61528516eb418740df355852f23425acc4d0656a