feat(3p): allow importer to pass stableNixpkgs and nixpkgs to use
This change makes readTree pass extra arguments to every tree member
which are passed when importing depot. It is however only possible to
pass _extra_ arguments, not to override ones already defined in
default.nix explicitly.

This is used in third_party/default.nix to accept nixpkgs and
stableNixpkgs as arguments which are used instead of the pinned versions
if given. This allows third party users of depot to import depot while
specifying a different nixpkgs version -- at their own risk as some
breakage is almost certain to ensue.

This is useful if you want to use (parts of) depot as a sub project
somewhere or import parts of depot into your system configuration
without having to depend on multiple versions of nixpkgs. Also this
allows for using depot without the need for builtins.fetchTarball during
evaluation which means that depot can be made friends with
restricted-eval and hydra if necessary.

Change-Id: I6fb0233230023521d1aa007ff6e5c9ca511494a6